requirements.txt
aws_cdk.aws_cloudwatch
aws_cdk.aws_cloudwatch_actions
aws-cdk.aws_lambda
aws-cdk.aws_logs
aws-cdk.aws_sns
aws-cdk.aws_sns_subscriptions
aws-cdk.aws_sqs
aws-cdk.core
requests